\name{.PVM.errors}
\alias{.PVM.errors}
\title{PVM Error Codes}
\description{PVM error codes}
\usage{
.PVM.errors
}
\details{
  Defines mappling of the character names to negative integers defined
  by PVM as error code.

  For internal use only.  Better error handling is being persued in rpvm.
}
\references{
  PVM documentation
}
\author{
  Na (Michael) Li \email{lina@u.washington.edu} and
  A.J. Rossini \email{rossini@u.washington.edu}
}
\keyword{utilities}
\keyword{connection}
\keyword{interface}


